Decoupled combos from BotD
