+++ src/parser/jobs/drg/index.js
@@ -1,6 +1,8 @@
 import About from './About'
+import BloodOfTheDragon from './BloodOfTheDragon'
 import Combos from './Combos'
 export default [
 	About,
+	BloodOfTheDragon,
 	Combos,
 ]
